backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
numpy
pandas
requests
httpx
lancedb
pdfplumber
python-pptx
python-docx
langchain-text-splitters
ollama
pyarrow
